Update 2013-05-13-phantom-js-headless-testing.markdown
As the comments point out, running the Selenium driver is unnecessary. I had followed some older setup instructions and it worked ok, so I hadn't tried it without Selenium. I thought the port settings were a little strange too, but it worked so I didn't worry about it. I've updated the post to reflect that change and to incorporate your advice about having Selenium available for use in other browsers.