Add note that Rest API testing requires the module-rest package
After working through the API testing documentation and attempting to
run the API tests, I encountered the following error:

Module REST is not installed.
Use Composer to install corresponding package:

composer require codeception/module-rest --dev

This change updates the documentation to cover this requirement, so that
it's not discovered when running the API tests.